终于知道bug在哪了 timer定时器的初始化设置函数uint16写成uint8了 导致一连串的bug

+++ lib/device/timer.h
@@ -6,6 +6,6 @@
 #define SYSTEM_TIMER_H
 
 #include "stdint.h"
-static  void freq_set(uint8_t counter_port,uint8_t counter_no,uint8_t rwl,uint8_t  counter_mode,uint8_t counter_value);
+static  void freq_set(uint8_t counter_port,uint8_t counter_no,uint8_t rwl,uint8_t  counter_mode,uint16_t counter_value);
 void timer_init();
 #endif //SYSTEM_TIMER_H
